% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_data}
\alias{summarize_data}
\title{Compute summary statistics, produce a tibble or data frame and a boxplot}
\usage{
summarize_data(df, x, y, na.rm = TRUE)
}
\arguments{
\item{df}{Tibble or data frame containing variables for computing summary statistics.}

\item{x}{A categorical variable by which rows are grouped.}

\item{y}{A numeric variable for computing summary statistics.}

\item{na.rm}{Gives the option to either remove or retain missing values.}
}
\value{
A tibble or data frame containing summary statistics and a boxplot.
}
\description{
This function computes summary statistics on a numeric variable
grouped by a categorical variable. The summary statistics include
mean, median, minimum, maximum, and count. It also
produces a boxplot.
}
\examples{
summarize_data(gapminder::gapminder, continent, lifeExp)
summarize_data(gapminder::gapminder, country, gdpPercap)
}
